Rename kintoclient into kinto_client.
