+++ kinto_client/exceptions.py
