+++ kinto_client/tests/config/kinto.ini
