+++ kinto_client/utils.py
