+++ tox.ini
@@ -4,11 +4,11 @@ envlist = py27,flake8
 [testenv:py27]
 commands =
     python --version
-    py.test kintoclient/tests/functional.py kintoclient/tests --cov-report term-missing --cov-fail-under 100 --cov kintoclient {posargs}
+    py.test kinto_client/tests/functional.py kinto_client/tests --cov-report term-missing --cov-fail-under 100 --cov kinto_client {posargs}
 deps = -rdev-requirements.txt
 install_command = pip install --pre {opts} {packages}
 
 [testenv:flake8]
-commands = flake8 kintoclient
+commands = flake8 kinto_client
 deps =
     flake8
